fix: move metrics package into pkg/
